pyproject.toml
[project]
name = "django-bootstrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cutie>=0.3.2",
    "python-slugify>=8.0.4",
]

[tool.uv.workspace]
members = [
    "project-slug",
]
